Use uint32 in place for ulong.

+++ src/common/device.h
@@ -243,10 +243,10 @@ public:
 
 #ifdef ENDIAN_IS_SMALL
 #define DEV_ID(a, b, c, d) \
-  (Device::ID(a + (ulong(b) << 8) + (ulong(c) << 16) + (ulong(d) << 24)))
+  (Device::ID(a + (uint32(b) << 8) + (uint32(c) << 16) + (uint32(d) << 24)))
 #else
 #define DEV_ID(a, b, c, d) \
-  (Device::ID(d + (ulong(c) << 8) + (ulong(b) << 16) + (ulong(a) << 24)))
+  (Device::ID(d + (uint32(c) << 8) + (uint32(b) << 16) + (uint32(a) << 24)))
 #endif
 
 // ---------------------------------------------------------------------------
