+++ src/common/soundsrc.h
@@ -10,7 +10,7 @@ typedef int16 Sample;
 class SoundSource {
  public:
   virtual int Get(Sample* dest, int size) = 0;
-  virtual ulong GetRate() = 0;
+  virtual uint32 GetRate() = 0;
   virtual int GetChannels() = 0;
   virtual int GetAvail() = 0;
 };
@@ -20,7 +20,7 @@ typedef int32 SampleL;
 class SoundSourceL {
  public:
   virtual int Get(SampleL* dest, int size) = 0;
-  virtual ulong GetRate() = 0;
+  virtual uint32 GetRate() = 0;
   virtual int GetChannels() = 0;
   virtual int GetAvail() = 0;
 };
